Bug: HotFix: Fix FCD subject count
The FCD was counted against sum of all of the subjects encountered
within that particular semester, except some students do not pick one
particular subject or they have electives, the subject sum naively
summed all of the subjects taken and just added them together.

This is a hotfix, It just replaced the subject sum directly to 800 for
semesters 1 to 7 and 600 for semester 8. Unsure how to proceed on
semester 8.